Wait for ksvc's separately from other deploys
It seems the K_SINK errors are red herrings and a "ready" ksvc won't
use any of its deploys crashlooping for that reason as long as there
are others that will work. So we ignore ksvc deploys when waiting for
the rest to be available.

Seems to work, and drogue.sh is now MUCH faster.