Drop clippy check for now
There are two problems with that. First, it doesn't run with the already
checked (pre-processed) files. So it will take more time to run this
in the "host".

Second, this only works for PRs coming from the same repository due to
a GitHub token/permission issue. So let's drop that for PRs, and may
enable this for the main branch instead.