Don't wrap the JSON config in "Data"
While the deprecation warning suggest to replace the old `.data()` call
`.app_data(web::Data())` also with the `JsonConfig` struct, the actual
code doesn't seem to require this, and actually check for both variants.

So the current code would work. However, the code first checks for the
unwrapped variant and then for the wrapped one.

> /// Extract payload config from app data. Check both `T` and `Data<T>`
> , in that order, and fall
> /// back to the default payload config.

So by not wrapping this, I think it becomes easier to read, but it
should be a bit faster to, as the first variant already is a hit.