ci: don't compile during check
The idea was to cache the intermediate files for all the projects in
the first (check) step, and then pull them from the cache in the
following steps (build and test).

However, the cache is using an ID from the hash of the Cargo.lock file
only. So only when the lock file changes, the intermediate changes will
be uploaded to the cache, and are available for the next job.

If the lock file is not changed, the check step wastes around 10 minutes
which doesn't get used, as the cache is not uploaded and not available
for the next step. So, we stop doing this.

If we would add the source files to the hash of the cache, then caching
this would work. However, this would invalidate the cache more often,
as the source changes much more often than the lock file.