Refactor actix-web dependency to service-api only
Initially, this is mostly re-exporting actix-web abstractions from
service-api, but it can evolve to be smarter and more suited to how
the drogue web apps use actix.

The biggest challenge was the actix-web::main proc macro, since it
generates code assuming actix-web is declared as a dep. Re-exporting
it from service-api required implementing a new proc-macro crate:
`macros`. Ugh.

The new macros crate could possibly justify its existence by moving
the `service-common/src/app.rs` stuff over to it, but that's for
another PR.

Removed actix-web dep from the following workspaces:
 service-common
 device-management-service
 console-backend
 user-auth-service
 coap-endpoint
 websocket-integration
 command-endpoint
 authentication-service
 integration-common (feature, too)
 admin-service (features, too)
 access-token-service
 database-common
 http-endpoint
 endpoint-common

I initially attempted to put the re-exports in service-common, but
because UserInformation is defined in service-api, its FromRequest
impl needed to live there, too. So I reasoned it was ok for
service-api to bring in the actix-web deps since most of the
workspaces depend on both service-api and service-common anyway. This
became a problem for console-frontend which targets wasm32, and some
of the actix deps won't compile there.

Therefore, we featurized the actix deps in service-api. Enabled by
default, because all but the console-frontend can build the actix
stuff.