First stab at porting to actix-web 4.0
The route macros now use absolute paths to reference ::actix_web::* so
this breaks our idea to re-export those things from a service-api
module.

So I think we either need to a) avoid using the macros or 2) declare a
dep on actix-web "4" in the workspace members that need the macros.

I took something of a hybrid approach here, but I ran into a type
mismatch error in the authentication-service wrapping the optional
prometheus middleware.