fix: bring back the detection that no services had been enabled
Before this was done by checking the number of threads. As there is now
one thread for each of the runtimes, this will no longer work, as
threads are created anyway, just exit right away.

Now the check evaluates if the exit was "too quick", and shows a warning
message.