fix: ensure the order of the connection lost event
When publishing a message, and immediately closing the connection,
it could happen that the connection lost event was sent before the
final data message.

This change ensures, using an R/W lock, that nothing is being published
after a disconnect message, and that the device state will only be
deleted once all pending publish requests have been processed.