fix: align generation field with k8s api
This change only incremenets the generation field when
a change of spec section has occurred. Internally there
is a revision field, which increments with every change.
However, that is not exposed, as the resourceVersion field
should be used for detecting any change. While the
generation field is intended to track changes of the
specification.

This fixes the "obverved generation" never being equal
to the "generation", as the status section update again
increments the generation field.