fix: prevent device state service from exiting due to misconfigured app
In case of an invalid 'publish' section in the application, the device
state service cannot publish. Not being able to publish means that the
device connection events cannot be sent. However we must send those
events. So the service will exit, and re-try.

However in the case a user misconfigured the spec section, the service
must not exit. So in this case, we log and drop the event.